lingua-tutor/pyproject.toml
[project]
name = "lingua-tutor"
version = "0.1.0"
description = "An AI-powered language learning assistant."
requires-python = ">=3.11, <3.12"
dependencies = [
    "loguru>=0.7.3",
    "openai-whisper",
    "pytest>=8.4.2",
    "torch==2.4.1",
    "torchaudio==2.4.1",
    "torchvision==0.19.1",
]

[tool.uv.pip]
extra-index-url = ["https://download.pytorch.org/whl/cu121"]
index-strategy = "unsafe-best-match"

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu121"

[[tool.uv.index]]
url = "https://pypi.org/simple"

[tool.uv.sources]
lingua-tutor = { path = "src" }
